Move WindowingStrategies to runners-core-construction
